pyproject.toml
[project]
name = "ai-usecase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.1.0",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.3",
    "pdfminer-six>=20251107",
    "pdfplumber>=0.11.8",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
